target-sh4: Introduce realizefn for SuperHCPU
To avoid moving cpu_reset() into an initfn for some targets, prepare a
realizefn with the designated signature but leave it non-static and call
it explicitly.
Do not declare a local env variable to avoid an unused variable
warning for *-user: qemu_init_vcpu() is a no-op macro there.

It is safe to reorder cpu_reset() and cpu_register() since there is no
overlap in the fields they touch.

Signed-off-by: Andreas Färber <afaerber@suse.de>