memory_mapping: Split out dump-guest-memory memory mapping code
The dump.c code currently requires CONFIG_HAVE_CORE_DUMP as well as
CONFIG_HAVE_GET_MEMORY_MAPPING. Some architectures will provide only the
non-paging case. Allow an architecture to provide dump support even when
CONFIG_HAVE_GET_MEMORY_MAPPING is not available. To do that, we split
out the common code and provide stub functions for the non-paging case.

If -p is specified on a target that doesn't support it, we will pass an
error to the calling code.

Signed-off-by: Ekaterina Tumanova <tumanova@linux.vnet.ibm.com>
Signed-off-by: Jens Freimann <jfrei@linux.vnet.ibm.com>
Signed-off-by: Andreas Färber <afaerber@suse.de>