memory_mapping: Cleanup qemu_get_guest_memory_mapping()
We relied on the CPUClass::get_memory_mapping() implementation being a
no-op when paging is disabled for that CPU. Therefore we can just as well
iterate over all CPUs to retrieve mappings.

This allows to use qemu_for_each_cpu() instead of open-coding CPU loops.

Pass any Error out into dump_init() and have it actually stop on errors.
Whether it is unsupported on a certain CPU can be checked by looking for
a NULL CPUClass::get_memory_mapping field.

Signed-off-by: Andreas Färber <afaerber@suse.de>